refactor: moves next-auth related module declarations to the global.d.ts
